compiler: improved function spec generation
Functions should have the right number of parameters, and they should be
of the right type when the type is _fixed_.

If the function has a single parameter which is a unit, then it is
dropped, as per the convention so far.

Also since atoms are mapped from polymorphic variant tags, we can turn
tagged tuples into polyvariants constructors, and print out the inferred
type specs.