parser: first successful parse of otp's binary.erl
A lot of changes going on here, but essentially the parser has been
extended to parse the entirety of the `binary.erl` module that has been
copied verbatim from the OTP codebase.

This example has driven changes in the AST as well, and we can now
represent a little more of the standard Erlang syntax:

* Types and Function Specs
* Comments on expressions, both single and multiple ones
* Operators (unary and binary infix ops)
* Try...catch
* If expressions
* Guards expressions

There's been some reformatting as well and some test results have been
updated, in particular the `operators.erl` test was useful to get the
operator support in place.

I have not exercised the printer a lot yet, but it has also been
extended to at least spit something out of the new AST version.