parser: preprocess module attribute values
Since we are parsing module attribute values with the much broader
`expr` production, we should preprocess the parsed values before
building up the actual final attribute.

We do this because the implicit syntax of exporting functions `f/a` gets
parsed as the application of the division function `erlang:'/'(f, a)`.
Unfortuante, but work-around-able.