compiler: separate ocaml->external ffi names
In order to continue building the standard library of FFIs to Erlang, I
had to add support to replace the name of a function for the appropriate
external name.

For example, the function `maps:new()` in Erlang would mean that we need
a function `Maps.new ()` in OCaml. This unfortunately won't work since
`new` is a reserved word.

However, by allowing the programmer to choose the final name to be used
by the FFI during the translation process, we can make the OCaml
function have a different name.

```ocaml
(* file: maps.ml *)
external make : unit -> map = "new"
```

Using `Maps.make ()` will compile to `maps:new()` in Erlang.

Works on #6