verify: fall back to parsetree verification
This is the first verification attempt that works, and it makes some
gross over genrealizations (such as all top-level structure items should
be marked as recursive, which isn't always necessary).

Lots of work to be done still but it is a start!

Progress towards #31

Thanks @Drup for the help :tada: