Fetch some dpeth of history
The current setup, may not name artifacts correctly given the way the
checkout it happening. This is generally good, but I am concerned that
the checkout will never have tags in it to give a valuable GITHUB_TAG
name. Trying the fetch-depth 1 option to handle the common scenario of a
branch off main (with tags) such that we should be able to get better
names.