ci: move to $GITHUB_ENV instead of ::set-env
* Echo to $GITHUB_ENV instead of ::set-env.

* Rely on `git` to get the tags instead of shell string manipulation.

* Use fetch-dept 1 to get better tag names.
The current setup, may not name artifacts correctly given the way the
checkout is happening. This is generally good, but I am concerned that
the checkout will never have tags in it to give a valuable `GITHUB_TAG`
name. Trying the fetch-depth 1 option to handle the common scenario of a
branch off main (with tags) such that we should be able to get better
names.

Closes #29