feat(sugarcane): correctly generate module hierarchy
I've rewritten the module name translation, following on Elixir
footsteps.

We're now using the new parametrizable parsetree, which can hold
location information on all nodes. This should make it easier to report
errors later on. With this comes a new set of tree helpers, and also a
new printer that I hope will be a lot easier to maintain.

Simplified a bunch of abstractions, and did a clearer split between the
Runner, the Caml_compiler, and the actual new compiler Sugarcane.

Also I've split the Sugarcane tests into the acceptance (old tests, most
of which we want to still pass), and a new test suite where I'll slowly
migrate tests as support is being added.